refactor: changing middlewares to folder
