+++ src/middlewares/index.js
@@ -1,1 +1,3 @@
-export * from './jwt-middleware'
+export * from './auth-middleware'
+export * from './error-handling-middleware'
+export * from './validation-middleware'
